moved session token to separate module
