+++ account_sdk/src/lib.rs
@@ -4,8 +4,10 @@ pub mod webauthn_signer;
 
 pub mod abigen;
 pub mod felt_ser;
+mod session_token;
+
 #[cfg(test)]
-mod tests;
+pub mod tests;
 
 use wasm_bindgen::prelude::*;
