+++ account_sdk/src/tests/mod.rs
@@ -1,8 +1,7 @@
-mod deployment_test;
+pub mod deployment_test;
 mod ecdsa_test;
 mod erc20_methods_test;
-mod runners;
+pub mod runners;
 mod runners_test;
-mod session_token;
 mod stark_pair_auth_test;
 mod webauthn_signer_test;
